# src/main/resources/OpenAPI.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8085",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/recipes/{rid}": {
      "get": {
        "tags": [
          "Get"
        ],
        "summary": "Get Recipe",
        "description": "Get a particular recipe by ID",
        "operationId": "getRecipe",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDTO"
                }
              }
            }
          },
          "200": {
            "description": "Recipe",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDTO"
                }
              }
            }
          },
          "500": {
            "description": "An internal error occurred",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDTO"
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Update"
        ],
        "summary": "Update Recipe",
        "description": "Update a particular recipe by ID",
        "operationId": "updateRecipe",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Recipes"
                }
              }
            }
          },
          "500": {
            "description": "An internal error occurred",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Recipes"
                }
              }
            }
          },
          "200": {
            "description": "Recipe is updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Recipes"
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Recipes"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Delete"
        ],
        "summary": "Delete Recipe",
        "description": "Delete a particular recipe by ID",
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "Recipe deleted",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "An internal error occurred",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/recipes": {
      "get": {
        "tags": [
          "Get"
        ],
        "summary": "Get Recipes",
        "description": "Get a list of recipes",
        "operationId": "getAllRecipes",
        "responses": {
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeDTO"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Recipe list",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeDTO"
                  }
                }
              }
            }
          },
          "500": {
            "description": "An internal error occurred",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Recipe list not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Post"
        ],
        "summary": "Add Recipe",
        "description": "Add a recipe",
        "operationId": "addRecipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "An internal error occurred",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "Add a recipe",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/filter": {
      "post": {
        "tags": [
          "Post"
        ],
        "summary": "Filter Recipe",
        "description": "Filter recipes ",
        "operationId": "getSearch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "List of filtered recipes",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recipes"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recipes"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RecipeDTO": {
        "type": "object",
        "properties": {
          "rid": {
            "type": "integer",
            "format": "int32"
          },
          "dish": {
            "type": "string"
          },
          "foodType": {
            "type": "string"
          },
          "numberOfServings": {
            "type": "integer",
            "format": "int32"
          },
          "ingredients": {
            "type": "string"
          },
          "instruction": {
            "type": "string"
          }
        }
      },
      "Recipes": {
        "type": "object",
        "properties": {
          "rid": {
            "type": "integer",
            "format": "int32"
          },
          "dish": {
            "type": "string"
          },
          "foodType": {
            "type": "string"
          },
          "numberOfServings": {
            "type": "integer",
            "format": "int32"
          },
          "ingredients": {
            "type": "string"
          },
          "instruction": {
            "type": "string"
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "foodType": {
            "type": "string"
          },
          "numberOfServings": {
            "type": "integer",
            "format": "int32"
          },
          "ingredients": {
            "type": "string"
          },
          "instruction": {
            "type": "string"
          }
        }
      }
    }
  }
}